pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "crop-disease"
version = "1.0.0"
description = "Crop disease classification using transfer learning with PyTorch"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Santosh Shinde"}]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchmetrics>=1.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "Pillow>=9.5.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
app = ["streamlit>=1.28.0", "plotly>=5.15.0"]
api = ["fastapi>=0.104.0", "uvicorn[standard]>=0.24.0", "python-multipart>=0.0.6"]
notebook = ["jupyter>=1.0.0", "ipykernel>=6.0.0"]
export = ["onnx>=1.14.0", "onnxscript>=0.1.0", "onnxsim>=0.4.0"]
test = ["pytest>=7.0.0", "httpx>=0.25.0"]
dev = [
    "crop-disease[app,api,notebook,test]",
    "ruff>=0.1.0",
]

[project.scripts]
export-model = "scripts.export_model:main"

[tool.setuptools.packages.find]
include = ["src*", "app*", "api*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "W", "UP"]
